requirements.txt
Adafruit-Blinka==8.69.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.15
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-pn532==2.4.6
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.86.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
pyftdi==0.57.1
pyserial==3.5
python-dotenv==1.2.1
pyusb==1.3.1
redis==7.1.0
requests==2.32.5
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
schedule==1.2.2
spotipy==2.25.2
sysv_ipc==1.2.0
typing_extensions==4.15.0
urllib3==2.6.3
